Fix non-sim weapon velocity calculation
Weapon velocity is converted to float tiles-per-second instead of the
simulated pixels-per-tick, but the calculation was off by a factor of
10. This didn't affect the simulation at all, but it did affect
decisions made in behavior trees to be wrong. The dodge node was off by
a lot, for example.

This broke during the weapon simulation update when it was converted to
work on pixels instead of floats for better accuracy.